Moved CoordinationService into it's own 'application' package
