+++ Core/src/org/sleuthkit/autopsy/application/CoordinationService.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.sleuthkit.autopsy.configuration;
+package org.sleuthkit.autopsy.application;
 
 import java.util.HashMap;
 import java.util.Map;
